// src/shared/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          created_at: string
          email: string
          id: string
          is_banned: boolean
          last_login_at: string | null
          manager_id: string | null
          name: string
          phone: string | null
          phone_verified_at: string | null
          remember_token: string | null
          role: string
          updated_at: string
          wallet_balance: number
        }
        Insert: {
          created_at?: string
          email: string
          id: string
          is_banned?: boolean
          last_login_at?: string | null
          manager_id?: string | null
          name: string
          phone?: string | null
          phone_verified_at?: string | null
          remember_token?: string | null
          role?: string
          updated_at?: string
          wallet_balance?: number
        }
        Update: {
          created_at?: string
          email?: string
          id?: string
          is_banned?: boolean
          last_login_at?: string | null
          manager_id?: string | null
          name?: string
          phone?: string | null
          phone_verified_at?: string | null
          remember_token?: string | null
          role?: string
          updated_at?: string
          wallet_balance?: number
        }
        Relationships: [
          {
            foreignKeyName: "profiles_manager_id_fkey"
            columns: ["manager_id"]
            isOneToOne: false
            referencedRelation: "profile_wallets_consolidated"
            referencedColumns: ["profile_id"]
          },
          {
            foreignKeyName: "profiles_manager_id_fkey"
            columns: ["manager_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}